Fix sonar failures

+++ hello-world-web/src/main/java/net/zoostar/hw/web/interceptor/GatewayAudit.java
@@ -11,7 +11,7 @@ public class GatewayAudit {
 
 	private long time;
 	
-	private double duration;
+	private long duration;
 	
 	private String endPoint;
